Lesson 2: Connect Sunshine to the Cloud
Learned about threads including the UI/Main thread and the background
thread, logging errors in the console at the five different levels,
giving permissions to the app such as access to internet, fetching data
from the OpenWeatherMap API, and parsing JSON data.  The app now
displays a week’s worth of weather forecasts for Toronto when the
refresh button is tapped! :)